% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sacva.R
\name{cva_capital}
\alias{cva_capital}
\title{CVA risk capital charge}
\usage{
cva_capital(cp_rating, cp_wam, cp_ead, cp_df = NULL, cp_hedge_wam = NULL,
  cp_hedge_notional = NULL, cp_hedge_df = NULL, cva_hedge_rating = NULL,
  cva_hedge_wam = NULL, cva_hedge_notional = NULL, cva_hedge_df = NULL)
}
\arguments{
\item{cp_rating}{the CP's long term credit rating which are mapped to
weighting factors in accordance with APS 180, Att. A, para. 18. The credit
ratings are defined in APS 120, Att. F, para. 13. with unrated
exposures mapped to credit rating 4. Must be a numeric vector with unrated
exposure ratings set to \code{NA}.}

\item{cp_wam}{the weighted average maturity of derivative cash flows with the
CP (\eqn{M}). Must be greater than zero.}

\item{cp_ead}{the exposure at default (\eqn{Exposure^total}) which must be no
less than 0.0}

\item{cp_df}{the discount factor (\eqn{D}) which must be no more than 1.0 or
\code{NULL}. Defaults to \code{NULL} which means that it is calculated from
\code{cp_wam} per the prescribed formula.}

\item{cp_hedge_wam}{the maturity of single name CDS cash flows referencing
the CPs (\eqn{M}). Must be greater than zero.}

\item{cp_hedge_notional}{notional amount of purchased single name CDS}

\item{cp_hedge_df}{the discount factor (\eqn{D}) which must be no more than
1.0 or \code{NULL}. Defaults to \code{NULL} which means that it is
calculated from \code{cp_hedge_wam} per the prescribed formula.}

\item{cva_hedge_rating}{the index CDS counterparty's long term external
credit rating. See \code{cp_rating} for further details}

\item{cva_hedge_wam}{maturity in years of the index CDS}

\item{cva_hedge_notional}{the notional amount of the index CDS}

\item{cva_hedge_df}{the discount factor (\eqn{D}) which must be no more than
1.0 or \code{NULL}. Defaults to \code{NULL} which means that it is
calculated from \code{wam} per the prescribed formula.}
}
\value{
the CVA capital charge for the counterparty or an error if an invalid
argument values are supplied.
}
\description{
This method determines a portfolio's CVA capital charge in accordance with
APS 180, Att. A, para. 17. You supply three sets of arguments to this
function: counterparty (CP) exposures, CP exposure single name hedges and
index CVA hedges. The arguments to this function are vectorised so that each
element of CP exposure and single name hedge arguments relate to one
counterparty (or netting set if a counterparty has more than one netting set)
and each element of the macro CVA hedge arguments relate to one CDS index. If
a vector of length greater than one is supplied for \code{cp_rating}, the
multiple counterparty CVA capital charge is calculated. In this case, if the
\code{cp_hedge_wam} or \code{cva_hedge_weight} arguments are not \code{NULL},
hedging benefits accrue. The lengths of each of the CP exposures and single
name hedge arguments must be consistent (or zero in case of \code{df = NULL})
while the index CVA hedge arguments must have the same length (or zero in
case of \code{df = NULL}).
}
\examples{
# Single counterparty
cva_capital(1, 2.3, 1e6, 0.8)
# Multiple counterparties
cva_capital(c(1, 2), c(1, 0.5), c(100, 8000))
# Multiple counterparties with hedges
cva_capital(c(2, 3), c(1, 0.5), c(100, 8000), NULL,
  c(1, 0.5), c(50, 1000), NULL, 1, 1, 2000, NULL)
}
\references{
\href{http://www.apra.gov.au/adi/PrudentialFramework/Documents/APS-180.pdf}{APS 180}
}
