% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/irb.R
\name{non_retail_capital}
\alias{non_retail_capital}
\title{Non-retail exposure capital requirements}
\usage{
non_retail_capital(pd, lgd, size, maturity, is_riskier_fi)
}
\arguments{
\item{pd}{the probability of default expressed as a decimal between 0 and 1
(exclusive). It is assumed that these are already floored to 0.0003 for
corporate and bank exposures (B.2).}

\item{lgd}{the loss given default expressed as a decimal. It is assumed that
any applicable floors have already been applied.}

\item{size}{the size of the consolidated group to which an obligor belongs
expressed in millions of dollars of sales}

\item{maturity}{the effective maturity expressed in years. It is assumed that
any applicable floors or ceilings have already been applied.}

\item{is_riskier_fi}{\code{TRUE} if an obligor is a regulated financial
institution with total assets greater than A$100bn or in an unregulated
financial institution. Otherwise is \code{FALSE}. See 113.B.77(a) and
113.B.77(b)}
}
\value{
a vector of capital requirement ratios subject to a floor of zero.
Note this assumes a Prudential Capital Ratio of 0.08
}
\description{
This calculates the capital requirements of corporate, bank and sovereign
exposures that are not specialised lending exposures subject to the slotting
approach. All inputs are vectors and whose values are recycled if necessary.
}
\examples{
non_retail_capital(0.05, 0.45, 3, 100, FALSE)
non_retail_capital(c(0.05, 0.70), 0.45, 3, 100, FALSE)
}
\references{
\href{https://www.legislation.gov.au/Details/F2012L02329}{APS 113}
}
