% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/irb.R
\name{slotting_capital}
\alias{slotting_capital}
\title{Slotted specialised lending exposure capital}
\usage{
slotting_capital(category)
}
\arguments{
\item{category}{the slotting category must be one of the following values:
\code{strong}, \code{good}, \code{satisfactory}, \code{weak}, \code{default}}
}
\value{
a vector of capital requirement ratios subject to a floor of zero.
Note this assumes a Prudential Capital Ratio of 0.08
}
\description{
This calculates the capital requirements of specialised lending exposures
subject to slotting.
}
\examples{
slotting_capital(c("strong", "weak"))
}
\seealso{
Other APS113 functions: \code{\link{non_retail_capital}},
  \code{\link{retail_capital}}
}
