% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/irb.R
\name{retail_capital}
\alias{retail_capital}
\title{Retail exposure capital requirements}
\usage{
retail_capital(pd, lgd, sub_class)
}
\arguments{
\item{pd}{the probability of default expressed as a decimal between 0 and 1
(exclusive). It is assumed that these are already floored to 0.0003 for
corporate and bank exposures (B.2).}

\item{lgd}{the loss given default expressed as a decimal. It is assumed that
any applicable floors have already been applied.}

\item{sub_class}{the sub-class of the retail IRB asset class. Can be one of
the following: \code{mortgage}, \code{qrr} (qualified revolving retail) and \code{other}
(e.g. personal lending and small business exposures)}
}
\value{
a vector of capital requirement ratios
}
\description{
This calculates the capital requirements of retail exposures including
residential mortgages, qualifying revolving retail and SME Retail (other).
All inputs are vectors and whose values are recycled if necessary.
}
\examples{
retail_capital(0.04, 0.20, "mortgage")
retail_capital(0.04, 0.20, "qrr")
retail_capital(0.04, 0.20, "other")
}
\references{
\href{https://www.legislation.gov.au/Details/F2012L02329}{APS 113}
}
\seealso{
Other APS113 functions: \code{\link{non_retail_capital}},
  \code{\link{slotting_capital}}
}
